// src/shared/generated/AppKitRuntimeTabProjectionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuntimeTabActivationPhaseRecord } from "./RuntimeTabActivationPhaseRecord";

export type AppKitRuntimeTabProjectionRecord = { tabId: string, name: string, phase: RuntimeTabActivationPhaseRecord, tabType: "role" | "workspace" | "popup", workspaceTemplate?: string, audioMuted: boolean, };